refactor(tevent): move to shared types
