+++ src/pages/Demos/CheckboxInput/CheckboxInput.tsx
@@ -1,8 +1,9 @@
 import MessageTypes from '@axa-fr/react-toolkit-form-core/dist/esm/MessageTypes';
 import CheckboxModes from '@axa-fr/react-toolkit-form-input-checkbox/dist/esm/CheckboxModes';
-import Layout, { TLayoutPage } from 'Layout';
+import Layout, { type TLayoutPage } from 'Layout';
 import LiveCode from 'shared/components/LiveCode';
-import { withEditor, useEditable, TEvent, Tknobs, EditorHeader, TReturnUseToggleEditor } from 'shared/components/Editor';
+import type { TEvent } from 'shared/types';
+import { withEditor, useEditable, type Tknobs, EditorHeader, type TReturnUseToggleEditor } from 'shared/components/Editor';
 import { TITLE_BAR, TITLE, DESIGN_SYSTEM_PATH, STORYBOOK_PATH, GITHUB_PACKAGE, NPM_NAME } from './constants';
 import knobs from './knobs.json';
