+++ src/pages/Demos/Layout/Layout.tsx
@@ -1,5 +1,6 @@
 import type { ClickEvent } from '@axa-fr/react-toolkit-core/dist/esm/withClickId.hoc';
-import { withEditor, useEditable, TEvent, EditorHeader, TReturnUseToggleEditor, Tknobs } from 'shared/components/Editor';
+import type { TEvent } from 'shared/types';
+import { withEditor, useEditable, EditorHeader, type TReturnUseToggleEditor, type Tknobs } from 'shared/components/Editor';
 import LiveCode from 'shared/components/LiveCode';
 import { useToggleModal } from 'shared/components/ModalCommon';
 import Layout from 'Layout';
