+++ src/pages/Demos/MembersNew/MembersNew.tsx
@@ -2,7 +2,7 @@ import Button from '@axa-fr/react-toolkit-button/dist/esm/index';
 import type { Control, FieldValues, UseFormReset } from 'react-hook-form';
 import { FieldTextInput, FieldSelectInput } from 'shared/components/form';
 import { Col, Row } from 'shared/components/Grid';
-import Layout, { TLayoutPage } from 'Layout';
+import Layout, { type TLayoutPage } from 'Layout';
 import {
   TITLE_BAR,
   TITLE,
