+++ src/pages/Demos/Popover/Popover.tsx
@@ -1,9 +1,10 @@
 import PopoverModes from '@axa-fr/react-toolkit-popover/dist/esm/PopoverModes';
 import PopoverPlacements from '@axa-fr/react-toolkit-popover/dist/esm/PopoverPlacements';
 import type { ClickEvent } from '@axa-fr/react-toolkit-core/dist/esm/index';
-import Layout, { TLayoutPage } from 'Layout';
+import Layout, { type TLayoutPage } from 'Layout';
+import type { TEvent } from 'shared/types';
 import LiveCode from 'shared/components/LiveCode';
-import { withEditor, useEditable, TEvent, Tknobs, EditorHeader, TReturnUseToggleEditor } from 'shared/components/Editor';
+import { withEditor, useEditable, type Tknobs, EditorHeader, type TReturnUseToggleEditor } from 'shared/components/Editor';
 import { TITLE_BAR, TITLE, DESIGN_SYSTEM_PATH, STORYBOOK_PATH, GITHUB_PACKAGE, NPM_NAME } from './constants';
 import knobs from './knobs.json';
