+++ src/pages/Home/Home.tsx
@@ -1,4 +1,4 @@
-import Layout, { TLayoutPage } from 'Layout';
+import Layout, { type TLayoutPage } from 'Layout';
 import { TITLE_BAR, TITLE } from './constants';
 import Galleries from './Galleries';
