+++ src/shared/components/Editor/index.ts
@@ -1,2 +1,2 @@
 export { withEditor, useEditable, EditorHeader } from './Editor';
-export type { TEvent, Tknobs, TReturnUseToggleEditor, TonChange } from './Editor';
+export type { Tknobs, TReturnUseToggleEditor, TonChange } from './Editor';
