+++ src/shared/components/form/form.helper.ts
@@ -1,4 +1,4 @@
 /* eslint-disable import/prefer-default-export */
-import type { TEvent } from 'shared/components/Editor';
+import type { TEvent } from 'shared/types';
 
 export const onChangeValue = (onChange: (value?: string) => void) => (e: TEvent) => onChange(e.value);
