+++ src/shared/types.d.ts
@@ -5,3 +5,10 @@ export type Tanomaly = {
   iconName?: string;
   code?: string | number;
 };
+
+export type TEvent = {
+  value?: string;
+  id?: string;
+  name?: string;
+  values?: string[];
+};
